+++ R/utilities.R
@@ -158,12 +158,19 @@ add_blank_row <- function(x, location="below", vars = NULL){
     else
       rw[1, i] <- rv
   }
+  
+  # Change copy position depending on location
+  cpos <- 1
+  if (location == "below")
+    cpos <- nrow(x)
 
   # Set page and page by values
   if ("..page_by" %in% names(x))
     rw[1, "..page_by"] <-  x[1, "..page_by"]
-  if ("..page" %in% names(x))
-    rw[1, "..page"] <-  x[1, "..page"]
+  if ("..page" %in% names(x)) {
+    # rw[1, "..page"] <-  x[1, "..page"]
+    rw[1, "..page"] <-  x[cpos, "..page"]
+  }
   
   # Add the blank row to the specified location.
   ret <- x
@@ -185,6 +192,9 @@ add_blank_row <- function(x, location="below", vars = NULL){
     rw2 <- rw
     rw$..blank <- "A"
     rw2$..blank <- "B"
+    
+    if ("..page" %in% names(x)) 
+      rw2[1, "..page"] <- x[nrow(x), "..page"]
     
     ret <- rbind(rw, ret, rw2)
   }